+++ backend/canisters/storage_index/impl/src/jobs/ensure_sufficient_active_buckets.rs
@@ -76,6 +76,7 @@ async fn create_bucket(args: CreateBucketArgs) {
 
     let result = create_and_install(
         None,
+        None,
         args.canister_wasm,
         args.init_canister_args,
         args.cycles_to_use,